+++ lib/src/metta/runner/environment.rs
