+++ lib/src/metta/runner/init.default.metta
