+++ lib/tests/metta.rs
@@ -3,8 +3,7 @@ use hyperon::metta::runner::stdlib::UNIT_ATOM;
 #[cfg(feature = "minimal")]
 use hyperon::metta::runner::stdlib2::UNIT_ATOM;
 use hyperon::metta::text::*;
-use hyperon::metta::runner::Metta;
-use hyperon::metta::environment::EnvBuilder;
+use hyperon::metta::runner::{Metta, EnvBuilder};
 
 #[test]
 fn test_reduce_higher_order() {
